Separate avatar from their items
Using `i_add` for these test cases appears to run into complications
with needing a container for liquids, and storage in inventory for the
item. These considerations get in the way of what we're testing, which
is simply the effects of using various items, and this can be done
without putting the item in the inventory.